// frontend/src/api/generated/types.gen.ts
export type LoginDto = {
    email: string;
    password: string;
};